+++ Kernel/Language/de.pm
@@ -25,7 +25,7 @@ sub Data {
     $Self->{DateFormatShort}     = '%D.%M.%Y';
     $Self->{DateInputFormat}     = '%D.%M.%Y';
     $Self->{DateInputFormatLong} = '%D.%M.%Y - %T';
-    $Self->{Completeness}        = 0.996943371943372;
+    $Self->{Completeness}        = 0.996944355098102;
 
     # csv separator
     $Self->{Separator}         = ';';
@@ -7949,6 +7949,10 @@ sub Data {
             'Authentifikationstyp für das Sendmail-Modul. Falls \'OAuth2-Token\' gewählt wurde, muss SendmailModule::OAuth2TokenConfigName ebenfalls konfiguriert werden.',
         'Name of the OAuth2 token configuration to use for sending mails if \'OAuth2 token\' was configured in SendmailModule::AuthenticationType.' =>
             'Name der OAuth2-Token-Konfiguration für den Versand von E-Mails, falls \'OAuth2-Token\' in SendmailModule::AuthenticationType konfiguriert wurde.',
+        'Limits the amount of emails which are sent within one batch. Disable this setting to have no limit.' =>
+            'Begrenzt die Anzahl der E-Mails, die in einem Durchgang versendet werden. Deaktivieren Sie die Einstellung, wenn Sie keine Limitierung benötigen.',
+        'Applies the limit defined in \'SendmailModule::RateLimit\' per sender address.' =>
+            'Wendet das mit \'SendmailModule::RateLimit\' festgelegte Limit je Absenderadresse an.',
         'Hosts that need a separate info about authentication method and token (instead of both in one line). Most commonly needed for Office 365 and Outlook.' =>
             'Hosts, die eine getrennte Info über Authentifizierungsmethode und Token benötigen (anstatt beides in einer Zeile). Wird am häufigsten für Office 365 und Outlook benötigt.',
         'This option enables a dropdown which will be displayed instead of the time unit input field.' =>